+++ src/math/random_generator.cpp
@@ -62,7 +62,7 @@ RandomGenerator::RandomGenerator() :
 int RandomGenerator::srand(int x)    {
   int x0 = x;
   while (x <= 0)                          // random seed of zero means
-    x = time(0) % RandomGenerator::rand_max; // randomize with time
+    x = static_cast<int>(time(0)) % RandomGenerator::rand_max; // randomize with time
 
   if (debug > 0)
     printf("==== srand(%10d) (%10d) rand_max=%x =====\n",
@@ -99,7 +99,7 @@ int RandomGenerator::rand(int u, int v) {
 float RandomGenerator::randf(float v) {
   float rv;
   do {
-    rv = (static_cast<float>(RandomGenerator::random()))/RandomGenerator::rand_max * v;
+    rv = (static_cast<float>(RandomGenerator::random())) / static_cast<float>(RandomGenerator::rand_max) * v;
   } while (rv >= v);                      // rounding might cause rv==v
 
   if (debug > 0)
